+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 ### [Added]
 
 * DeleteAction : suppression des upload, stored_data, configuration et offering dans un workflow
+* Offering :
+    * ajout fonction Offering.api_synchronize() : synchronisation de l'offre avec la configuration
+    * ajout de la fonction Offering.get_url() : récupération de la liste des urls d'une offre
+    * ajout de l'action `SynchronizeOfferingAction` : synchronisation de l'offre avec la configuration depuis un workflow
 
 ### [Changed]